Allow NSNumber to be used instead of string for conversion to dates.
As some JSON feeds will not quote their numbers, they get interpreted
as NSNumber by NSJSONSerialization.

It uses the simple constructor for NSDate with unix timestamp from 1970, although the code is really simple, it allows us to keep the flow with using Realm+JSON just because we are using a unix epoch.